# [sdk-metrics] Improve the perf of histogram snapshots