--- a/src/OpenTelemetry/Metrics/MetricPoint.cs
+++ b/src/OpenTelemetry/Metrics/MetricPoint.cs
@@ -915,16 +915,7 @@ public struct MetricPoint
                         histogramBuckets.RunningSum = 0;
                     }
 
-                    Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
-
-                    for (int i = 0; i < histogramBuckets.RunningBucketCounts!.Length; i++)
-                    {
-                        histogramBuckets.SnapshotBucketCounts[i] = histogramBuckets.RunningBucketCounts[i];
-                        if (outputDelta)
-                        {
-                            histogramBuckets.RunningBucketCounts[i] = 0;
-                        }
-                    }
+                    histogramBuckets.Snapshot(outputDelta);
 
                     this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect(this.Tags, outputDelta);
 
@@ -979,16 +970,7 @@ public struct MetricPoint
                         histogramBuckets.RunningMax = double.NegativeInfinity;
                     }
 
-                    Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
-
-                    for (int i = 0; i < histogramBuckets.RunningBucketCounts!.Length; i++)
-                    {
-                        histogramBuckets.SnapshotBucketCounts[i] = histogramBuckets.RunningBucketCounts[i];
-                        if (outputDelta)
-                        {
-                            histogramBuckets.RunningBucketCounts[i] = 0;
-                        }
-                    }
+                    histogramBuckets.Snapshot(outputDelta);
 
                     this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect(this.Tags, outputDelta);
                     this.MetricPointStatus = MetricPointStatus.NoCollectPending;
@@ -1181,16 +1163,7 @@ public struct MetricPoint
                         histogramBuckets.RunningSum = 0;
                     }
 
-                    Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
-
-                    for (int i = 0; i < histogramBuckets.RunningBucketCounts!.Length; i++)
-                    {
-                        histogramBuckets.SnapshotBucketCounts[i] = histogramBuckets.RunningBucketCounts[i];
-                        if (outputDelta)
-                        {
-                            histogramBuckets.RunningBucketCounts[i] = 0;
-                        }
-                    }
+                    histogramBuckets.Snapshot(outputDelta);
 
                     this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect(this.Tags, outputDelta);
 
@@ -1247,16 +1220,7 @@ public struct MetricPoint
                         histogramBuckets.RunningMax = double.NegativeInfinity;
                     }
 
-                    Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
-
-                    for (int i = 0; i < histogramBuckets.RunningBucketCounts!.Length; i++)
-                    {
-                        histogramBuckets.SnapshotBucketCounts[i] = histogramBuckets.RunningBucketCounts[i];
-                        if (outputDelta)
-                        {
-                            histogramBuckets.RunningBucketCounts[i] = 0;
-                        }
-                    }
+                    histogramBuckets.Snapshot(outputDelta);
 
                     this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect(this.Tags, outputDelta);
                     this.MetricPointStatus = MetricPointStatus.NoCollectPending;
@@ -1429,15 +1393,13 @@ public struct MetricPoint
 
         int i = histogramBuckets!.FindBucketIndex(number);
 
-        Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
-
         AcquireLock(ref histogramBuckets.IsCriticalSectionOccupied);
 
         unchecked
         {
             this.runningValue.AsLong++;
             histogramBuckets.RunningSum += number;
-            histogramBuckets.RunningBucketCounts![i]++;
+            histogramBuckets.BucketCounts[i].RunningValue++;
 
             if (reportExemplar && isSampled)
             {
@@ -1462,13 +1424,11 @@ public struct MetricPoint
 
         AcquireLock(ref histogramBuckets.IsCriticalSectionOccupied);
 
-        Debug.Assert(histogramBuckets.RunningBucketCounts != null, "histogramBuckets.RunningBucketCounts was null");
-
         unchecked
         {
             this.runningValue.AsLong++;
             histogramBuckets.RunningSum += number;
-            histogramBuckets.RunningBucketCounts![i]++;
+            histogramBuckets.BucketCounts[i].RunningValue++;
 
             if (reportExemplar && isSampled)
             {