## Changes

* For histograms with buckets switch from two arrays (one for running values and one for snapshot values) to a single array containing a structure with both values. This speeds up the snapshot logic because we can now elide bounds checking. That is important because when it comes to histograms the snapshot blocks the recording of values. Making it faster reduces contention which increases throughput when a collection occurs.

## Benchmarks

| Method      | OutputDelta | Mean       | Error     | StdDev    |
|------------ |------------ |-----------:|----------:|----------:|
| UpdateOld   | False       |  0.2130 ns | 0.0222 ns | 0.0208 ns |
| UpdateNew   | False       |  0.1810 ns | 0.0264 ns | 0.0247 ns |
| SnapshotOld | False       |  9.8634 ns | 0.1283 ns | 0.1200 ns |
| SnapshotNew | False       |  4.0386 ns | 0.0894 ns | 0.0836 ns |
| UpdateOld   | True        |  0.1948 ns | 0.0247 ns | 0.0231 ns |
| UpdateNew   | True        |  0.1815 ns | 0.0120 ns | 0.0107 ns |
| SnapshotOld | True        | 10.7866 ns | 0.2127 ns | 0.1885 ns |
| SnapshotNew | True        |  7.3044 ns | 0.1561 ns | 0.1603 ns |

<details>
<summary>Code</summary>

```csharp
public class HistogramSnapshotBenchmarks
{
    private readonly HistogramBucketValues[] bucketCounts = new HistogramBucketValues[16];
    private readonly long[] runningBucketCounts = new long[16];
    private readonly long[] snapshotBucketCounts = new long[16];

    [Params(true, false)]
    public bool OutputDelta { get; set; }

    [Benchmark]
    public long UpdateOld()
    {
        return this.runningBucketCounts[0]++;
    }

    [Benchmark]
    public long UpdateNew()
    {
        return this.bucketCounts[0].RunningValue++;
    }

    [Benchmark]
    public void SnapshotOld()
    {
        for (int i = 0; i < this.runningBucketCounts.Length; i++)
        {
            this.snapshotBucketCounts[i] = this.runningBucketCounts[i];
            if (this.OutputDelta)
            {
                this.runningBucketCounts[i] = 0;
            }
        }
    }

    [Benchmark]
    public void SnapshotNew()
    {
        var bucketCounts = this.bucketCounts;

        if (this.OutputDelta)
        {
            for (int i = 0; i < bucketCounts.Length; i++)
            {
                ref var values = ref bucketCounts[i];
                ref var running = ref values.RunningValue;
                values.SnapshotValue = running;
                running = 0L;
            }
        }
        else
        {
            for (int i = 0; i < bucketCounts.Length; i++)
            {
                ref var values = ref bucketCounts[i];
                values.SnapshotValue = values.RunningValue;
            }
        }
    }

    internal struct HistogramBucketValues
    {
        public long RunningValue;
        public long SnapshotValue;
    }
}
```
</details>
## Merge requirement checklist

* [X] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-dotnet/blob/main/CONTRIBUTING.md) guidelines followed (license requirements, nullable enabled, static analysis, etc.)